Fix return values of recv/send functions
In the current headers the recv and send class functions return an int.
However, according to the man pages they should return a ssize_t.

Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #17

+++ include/sys/socket.h
@@ -90,13 +90,13 @@ int setsockopt(int s, int level, int optname, const void *optval,
 	       socklen_t optlen);
 int connect(int s, const struct sockaddr *name, socklen_t namelen);
 int listen(int s, int backlog);
-int recv(int s, void *mem, size_t len, int flags);
-int recvfrom(int s, void *mem, size_t len, int flags,
+ssize_t recv(int s, void *mem, size_t len, int flags);
+ssize_t recvfrom(int s, void *mem, size_t len, int flags,
 	     struct sockaddr *from, socklen_t *fromlen);
-int recvmsg(int s, struct msghdr *msg, int flags);
-int send(int s, const void *dataptr, size_t size, int flags);
-int sendmsg(int s, const struct msghdr *message, int flags);
-int sendto(int s, const void *dataptr, size_t size, int flags,
+ssize_t recvmsg(int s, struct msghdr *msg, int flags);
+ssize_t send(int s, const void *dataptr, size_t size, int flags);
+ssize_t sendmsg(int s, const struct msghdr *message, int flags);
+ssize_t sendto(int s, const void *dataptr, size_t size, int flags,
 	   const struct sockaddr *to, socklen_t tolen);
 int socketpair(int domain, int type, int protocol, int sv[2]);
